Move ColorListPicker into duotone-picker folder
