+++ packages/components/src/duotone-picker/color-list-picker/style.scss
